moves ScenarioList to its own module
also adds ScenarioList in init file of SLAMpy to give access to it when importing SLAMpy
