+++ SLAMpy/__init__.py
@@ -1,3 +1,4 @@
 
 
 from .scenario import ScenarioV3
+from .scenariolist import ScenarioList
